removed spirekey and added the rounded variant for chainweaver alpha icon
