+++ .changeset/thirty-cameras-run.md
@@ -1,0 +1,5 @@
+---
+'@kadena/kode-icons': patch
+---
+
+updated icons
