--- packages/libs/kode-icons/src/product/SpireKeyKdacolorLight.tsx
@@ -1,63 +1,0 @@
-import * as React from 'react';
-import type { SVGProps } from 'react';
-interface SVGRProps {
-  title?: string;
-  titleId?: string;
-}
-const SpireKeyKdacolorLight = (
-  { title, titleId, ...props }: SVGProps<SVGSVGElement> & SVGRProps,
-) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    data-style="kdacolor"
-    viewBox="0 0 264 64"
-    fontSize="1.5em"
-    fill="currentColor"
-    height="1em"
-    aria-labelledby={titleId}
-    {...props}
-  >
-    {title ? <title id={titleId}>{title}</title> : null}
-    <path
-      fill="#0B1D2E"
-      d="M98.004 16.695q.58-2.161 2.282-8.656h2.837q.567 2.16 2.282 8.656h-1.678q-.122-.471-.471-1.908h-3.091l-.47 1.908zm2.535-3.417h2.33q-.266-1.05-1.05-4.213h-.229q-.253 1.05-1.051 4.213M90.326 16.695V8.039h3.103q.422 1.883 1.714 7.545h.217V8.039h1.606v8.656h-3.103q-.422-1.884-1.714-7.546h-.217v7.546zM83.48 16.695V8.039h5.554v1.484H85.11v2.065h3.586v1.485h-3.585v2.137h4.008v1.485zM74.994 16.695V8.039h3.561q1.738 0 2.656.88.906.894.906 2.633v1.63q0 1.75-.906 2.631-.918.882-2.656.882zm1.63-1.485h1.955q.99 0 1.449-.52.459-.518.459-1.46v-1.727q0-.953-.46-1.46-.458-.52-1.448-.52h-1.955zM66.809 16.695q.58-2.161 2.281-8.656h2.837l2.282 8.656h-1.678q-.12-.471-.47-1.908H68.97q-.12.483-.471 1.908zm2.535-3.417h2.33q-.266-1.05-1.05-4.213h-.23q-.254 1.05-1.05 4.213M59.674 16.695V8.039h1.63v3.44h.217l2.813-3.44h2.088q-.905 1.074-3.61 4.261l3.73 4.395h-2.148q-.724-.882-2.873-3.514h-.218v3.514zM252.81 46.531V28.796h4.052v2.956h.185q.485-1.536 1.663-2.367 1.19-.843 2.713-.843.346 0 .774.035a5 5 0 0 1 .727.08v3.845q-.265-.092-.843-.161a8 8 0 0 0-1.097-.081q-1.143 0-2.055.497-.9.484-1.42 1.35t-.52 1.998V46.53zM241.387 46.878q-2.667 0-4.607-1.109-1.927-1.12-2.967-3.163-1.04-2.055-1.039-4.838 0-2.737 1.039-4.804 1.05-2.078 2.933-3.233 1.881-1.166 4.422-1.166a9 9 0 0 1 3.094.531q1.467.52 2.587 1.617 1.131 1.096 1.778 2.794.646 1.686.646 4.018v1.282h-14.536v-2.818h10.53q-.012-1.2-.52-2.136a3.77 3.77 0 0 0-1.42-1.49q-.9-.541-2.101-.542-1.282 0-2.252.624a4.3 4.3 0 0 0-1.512 1.616 4.64 4.64 0 0 0-.543 2.182v2.46q0 1.547.566 2.655a4.04 4.04 0 0 0 1.581 1.686q1.017.578 2.379.578.912 0 1.651-.254.74-.266 1.282-.774.543-.508.82-1.259l3.902.44a6.1 6.1 0 0 1-1.408 2.701q-1.028 1.143-2.633 1.778-1.605.624-3.672.624M231.247 28.796l-6.316 17.735h-4.619l-6.316-17.735h4.457l4.076 13.175h.185l4.087-13.175zM202.473 46.89q-1.686 0-3.037-.601a4.96 4.96 0 0 1-2.124-1.801q-.774-1.189-.774-2.933 0-1.5.555-2.483a4.3 4.3 0 0 1 1.512-1.57 7.6 7.6 0 0 1 2.159-.889 17.6 17.6 0 0 1 2.506-.45 99 99 0 0 0 2.529-.289q.97-.137 1.408-.416.45-.288.451-.889V34.5q0-1.305-.774-2.02t-2.229-.716q-1.535 0-2.436.67-.89.669-1.201 1.581l-3.902-.554q.462-1.617 1.524-2.702 1.062-1.097 2.598-1.64 1.535-.554 3.394-.554 1.282 0 2.552.3t2.321.993q1.05.681 1.686 1.86.646 1.176.646 2.944v11.87h-4.018v-2.437h-.138a5.1 5.1 0 0 1-1.074 1.386q-.681.634-1.721 1.027-1.027.381-2.413.381m1.086-3.072q1.258 0 2.182-.497.924-.507 1.42-1.339a3.4 3.4 0 0 0 .508-1.813v-2.09q-.196.162-.67.3a10 10 0 0 1-1.039.243q-.577.104-1.143.185l-.981.138a7.3 7.3 0 0 0-1.675.416q-.738.29-1.166.808-.427.509-.427 1.317 0 1.154.843 1.743t2.148.589M185.947 46.878q-2.666 0-4.607-1.109-1.928-1.12-2.967-3.163-1.04-2.055-1.039-4.838 0-2.737 1.039-4.804 1.05-2.078 2.933-3.233 1.882-1.166 4.422-1.166 1.64 0 3.094.531 1.467.52 2.587 1.617 1.131 1.096 1.778 2.794.647 1.686.647 4.018v1.282h-14.537v-2.818h10.53q-.012-1.2-.52-2.136a3.77 3.77 0 0 0-1.42-1.49q-.9-.541-2.101-.542-1.282 0-2.252.624a4.3 4.3 0 0 0-1.512 1.616 4.64 4.64 0 0 0-.543 2.182v2.46q0 1.547.566 2.655a4.03 4.03 0 0 0 1.582 1.686q1.016.578 2.378.578.912 0 1.651-.254.74-.266 1.282-.774.543-.508.82-1.259l3.902.44a6.1 6.1 0 0 1-1.408 2.701q-1.028 1.143-2.633 1.778-1.605.624-3.672.624M155.052 46.531l-5.011-17.735h4.26l3.118 12.47h.162l3.186-12.47h4.215l3.187 12.4h.173l3.071-12.4h4.272l-5.022 17.735h-4.353l-3.326-11.985h-.242l-3.326 11.985zM135.758 36.14V46.53h-4.18V28.796h3.995v3.014h.208a5.07 5.07 0 0 1 1.951-2.367q1.351-.878 3.337-.878 1.836 0 3.199.785 1.374.786 2.124 2.275.762 1.49.751 3.614V46.53h-4.18V35.886q0-1.779-.924-2.783-.912-1.005-2.528-1.005-1.097 0-1.952.485a3.37 3.37 0 0 0-1.328 1.374q-.473.9-.473 2.183M123.1 46.531V28.796h4.18v17.735zm2.102-20.252q-.993 0-1.709-.658-.716-.67-.716-1.605 0-.947.716-1.605a2.42 2.42 0 0 1 1.709-.67q1.004 0 1.708.67.716.659.716 1.605 0 .934-.716 1.605-.704.658-1.708.658M109.533 46.89q-1.686 0-3.036-.601a4.95 4.95 0 0 1-2.125-1.801q-.774-1.189-.774-2.933 0-1.5.555-2.483a4.3 4.3 0 0 1 1.512-1.57 7.6 7.6 0 0 1 2.159-.889 17.6 17.6 0 0 1 2.506-.45 99 99 0 0 0 2.529-.289q.97-.137 1.408-.416.45-.288.451-.889V34.5q0-1.305-.774-2.02t-2.228-.716q-1.536 0-2.437.67-.888.669-1.201 1.581l-3.902-.554q.462-1.617 1.524-2.702 1.062-1.097 2.598-1.64 1.536-.554 3.394-.554 1.282 0 2.552.3t2.321.993q1.05.681 1.686 1.86.646 1.176.646 2.944v11.87h-4.018v-2.437h-.138a5.1 5.1 0 0 1-1.074 1.386q-.681.634-1.721 1.027-1.027.381-2.413.381m1.086-3.072q1.258 0 2.182-.497.924-.507 1.42-1.339a3.4 3.4 0 0 0 .508-1.813v-2.09q-.196.162-.67.3a9 9 0 0 1-1.039.243q-.578.104-1.143.185l-.981.138a7.3 7.3 0 0 0-1.675.416q-.738.29-1.166.808-.427.509-.427 1.317 0 1.154.843 1.743t2.148.589M88.669 36.14V46.53h-4.18V22.884h4.087v8.926h.208q.624-1.501 1.928-2.367 1.317-.878 3.349-.878 1.848 0 3.221.774 1.374.774 2.125 2.263.762 1.49.762 3.637V46.53h-4.18V35.886q0-1.79-.924-2.783-.912-1.005-2.563-1.005-1.108 0-1.986.485a3.4 3.4 0 0 0-1.362 1.374q-.485.9-.485 2.182M80.75 30.863h-4.317a5.4 5.4 0 0 0-.682-1.882 5.1 5.1 0 0 0-1.235-1.409 5.2 5.2 0 0 0-1.686-.866q-.935-.3-2.02-.3-1.929 0-3.418.97-1.489.959-2.333 2.817-.843 1.848-.843 4.515 0 2.714.843 4.572.855 1.848 2.333 2.794 1.49.936 3.406.936 1.062 0 1.986-.277a5.4 5.4 0 0 0 1.674-.843q.75-.555 1.259-1.363a5.15 5.15 0 0 0 .716-1.847l4.318.023a9.3 9.3 0 0 1-1.05 3.164 9.3 9.3 0 0 1-2.09 2.61q-1.294 1.118-3.026 1.754-1.731.624-3.845.624-3.117 0-5.565-1.444t-3.857-4.168-1.408-6.535q0-3.822 1.42-6.535 1.42-2.725 3.868-4.169t5.542-1.443q1.975 0 3.672.554 1.697.555 3.025 1.628a8.7 8.7 0 0 1 2.182 2.61q.867 1.536 1.132 3.51M257.168 61.724h-.86l3.27-8.99h.878l3.271 8.99h-.861l-2.818-7.924h-.061zm.619-3.446h4.46v.737h-4.46zM248.029 61.724v-8.99h.82v4.121h5.101v-4.122h.821v8.99h-.821v-4.13h-5.101v4.13zM240.302 61.724v-8.99h2.902q.957 0 1.598.368.645.365.97.992.33.628.329 1.413 0 .786-.324 1.419a2.4 2.4 0 0 1-.966.996q-.641.364-1.594.364h-2.282v-.737h2.26q.707 0 1.173-.264.465-.267.693-.728.233-.461.233-1.05 0-.587-.233-1.049a1.67 1.67 0 0 0-.698-.724q-.465-.264-1.181-.263h-2.059v8.253zM233.349 61.724v-8.99h.821v8.252h4.284v.738zM225.25 61.724h-.861l3.271-8.99h.878l3.27 8.99h-.86l-2.819-7.924h-.061zm.619-3.446h4.46v.737h-4.46zM216.711 61.847a3.5 3.5 0 0 1-1.484-.299 2.57 2.57 0 0 1-1.036-.839 2.2 2.2 0 0 1-.403-1.233h.829a1.5 1.5 0 0 0 .303.86q.28.365.746.571a2.6 2.6 0 0 0 1.036.202q.624 0 1.098-.233a1.83 1.83 0 0 0 .746-.636 1.6 1.6 0 0 0 .272-.922q0-.544-.276-.961a1.8 1.8 0 0 0-.795-.65q-.513-.233-1.229-.233h-.527v-.737h.527a2.4 2.4 0 0 0 1.032-.211q.452-.215.711-.601a1.6 1.6 0 0 0 .259-.91q0-.495-.224-.873a1.56 1.56 0 0 0-.632-.592 2 2 0 0 0-.944-.215q-.522 0-.957.201-.435.202-.702.571-.264.369-.281.874h-.795q.018-.703.386-1.234.373-.535.988-.834a3.1 3.1 0 0 1 1.37-.303q.795 0 1.374.325a2.33 2.33 0 0 1 1.216 2.046q0 .767-.417 1.317a2.15 2.15 0 0 1-1.124.755v.057q.86.171 1.352.764.496.588.496 1.444 0 .72-.377 1.295a2.65 2.65 0 0 1-1.036.904q-.654.33-1.502.33M211.488 61.79a.62.62 0 0 1-.447-.185.62.62 0 0 1-.185-.448.6.6 0 0 1 .185-.447.62.62 0 0 1 .447-.185.6.6 0 0 1 .448.184.6.6 0 0 1 .184.448q0 .172-.087.316a.65.65 0 0 1-.229.233.6.6 0 0 1-.316.083M204.123 52.733l2.81 7.924h.079l2.809-7.924h.861l-3.271 8.99h-.878l-3.27-8.99z"
-    />
-    <path
-      fill="#4A9079"
-      d="M24.42 27.08v-6.77l5.863-3.386 5.866 3.386v6.77l-5.866 3.387zM12.69 27.08v-6.77l5.864-3.385 5.866 3.385v6.77l-5.866 3.388z"
-    />
-    <path
-      fill="#4A9079"
-      d="M18.56 30.464v-6.77l5.863-3.386 5.865 3.386v6.77l-5.865 3.387z"
-    />
-    <path
-      fill="#4A9079"
-      d="M18.555 37.237v-6.771l5.864-3.385 5.865 3.385v6.77l-5.865 3.388z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M.966 40.63v-6.77l5.864-3.386 5.865 3.385v6.771L6.83 44.017zM.966 27.085v-6.77L6.83 16.93l5.865 3.385v6.77L6.83 30.472z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M.961 20.314v-6.77l5.864-3.386 5.865 3.386v6.77l-5.865 3.387zM18.557 10.152v-6.77l5.864-3.386 5.866 3.385v6.771l-5.866 3.387z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M6.83 16.926v-6.77l5.864-3.386 5.865 3.386v6.77l-5.865 3.387zM12.694 47.4v-6.77l5.863-3.385 5.866 3.385v6.77l-5.866 3.388z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M18.554 50.78v-6.771l5.864-3.386 5.865 3.386v6.77l-5.865 3.387zM30.285 16.923v-6.771l5.864-3.385 5.865 3.385v6.77L36.15 20.31zM36.149 40.625v-6.771l5.864-3.385 5.865 3.385v6.77l-5.865 3.388z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M24.421 47.399v-6.77l5.864-3.386 5.865 3.385v6.77l-5.865 3.388zM36.149 27.08v-6.77l5.864-3.386 5.865 3.385v6.771l-5.865 3.387z"
-    />
-    <path
-      fill="#0B1D2E"
-      d="M36.148 20.308v-6.77l5.864-3.386 5.865 3.385v6.77l-5.865 3.388z"
-    />
-  </svg>
-);
-export default SpireKeyKdacolorLight;
