+++ packages/libs/kode-icons/src/product/index.ts
@@ -2,6 +2,8 @@ export { default as ChainweaverAlphaKdacolorDark } from './ChainweaverAlphaKdaco
 export { default as ChainweaverAlphaKdacolorLight } from './ChainweaverAlphaKdacolorLight';
 export { default as ChainweaverAlphaLogoKdacolorDark } from './ChainweaverAlphaLogoKdacolorDark';
 export { default as ChainweaverAlphaLogoKdacolorLight } from './ChainweaverAlphaLogoKdacolorLight';
+export { default as ChainweaverAlphaRoundedKdacolorDark } from './ChainweaverAlphaRoundedKdacolorDark';
+export { default as ChainweaverAlphaRoundedKdacolorLight } from './ChainweaverAlphaRoundedKdacolorLight';
 export { default as DocsKdacolorDark } from './DocsKdacolorDark';
 export { default as DocsKdacolorLight } from './DocsKdacolorLight';
 export { default as ExplorerKdacolorDark } from './ExplorerKdacolorDark';
@@ -12,9 +14,5 @@ export { default as MarmaladeV2KdacolorDark } from './MarmaladeV2KdacolorDark';
 export { default as MarmaladeV2KdacolorLight } from './MarmaladeV2KdacolorLight';
 export { default as ProofOfUsKdacolorDark } from './ProofOfUsKdacolorDark';
 export { default as ProofOfUsKdacolorLight } from './ProofOfUsKdacolorLight';
-export { default as SpireKeyKdacolorDark } from './SpireKeyKdacolorDark';
-export { default as SpireKeyKdacolorLight } from './SpireKeyKdacolorLight';
-export { default as SpireKeyLogoKdacolorDark } from './SpireKeyLogoKdacolorDark';
-export { default as SpireKeyLogoKdacolorLight } from './SpireKeyLogoKdacolorLight';
 export { default as ToolsKdacolorDark } from './ToolsKdacolorDark';
 export { default as ToolsKdacolorLight } from './ToolsKdacolorLight';
